Refactor Code for Unknown Final Four Matchups
Logic is added to flexibly determine which regions play each other in
the Final Four. There is currently not a solution to determine which
Final Four team from which region plays one another. The debugging
methods that print the data and force the data to be in a particular
state are modified accordingly.

The client-side logic in the Handlebars page that renders the content
for the dialog box is moved to a separate JS file. More edge cases are
handled related to the user switching between 'My Picks' and 'Results'
when viewing the bracket.