Fix Bug in Saving and Displaying Bracket
The logic for saving the bracket and displaying the saved bracket is
modified in order to account for the special cases of teams in the
national bracket.

To account for the delay in saving the bracket when the user clicks on
the button, the button's label is changed to disable the button and
alert the user that it is in the process of saving.